fix: moved jwt functions to server.py
fixes error 500 when user tries to login
