--- app/jwt.py
@@ -1,17 +1,0 @@
-from fastapi import Request
-from fastapi.responses import JSONResponse
-from fastapi_jwt_auth import AuthJWT
-from fastapi_jwt_auth.exceptions import AuthJWTException
-
-from app.server import app
-from app.config import CONFIG
-
-
-@AuthJWT.load_config
-def get_config():
-    return CONFIG
-
-
-@app.exception_handler(AuthJWTException)
-def jwt_exception_handler(request: Request, exc: AuthJWTException):
-    return JSONResponse(status_code=exc.status_code, content={"detail": exc.message})
